Make refs normative when needed, update QUIC and CDDL refs
Several IETF standards on which the Open Screen Protocol is based appeared in
the list of informative references instead of in the list of normative
references.

The QUIC entry in the local biblio had the wrong title, and linked to a now
outdated draft. The update lists referenced sections with their title instead
of the section number (more explicit) in the list of custom definitions. The
spec referenced the `APPLICATION_CLOSE` frame section in draft-16, which I
believe got merged into the `CONNECTION_CLOSE` frame section in draft-23.

Also, the CDDL spec has now been published as RFC8610.