Switch build to GitHub action using spec-prod
This creates a GitHub action that replaces the Travis CI worfklow for
re-generating the specification and updating the `gh-pages` branch.

The action uses w3c/spec-prod under the hoods:
https://github.com/w3c/spec-prod/#spec-prod

w3c/spec-prod takes care of everything and also validates the spec (markup and
broken hyperlinks).

Good thing about this approach is that no specific secret tokens are needed,
since GitHub actions are allowed to push to the gh-pages branch. Also, no need
to maintain a complex deploy script.