Convert HTMLMediaElements links to autolinks
This update converts all explicit references to the HTML spec to Bikeshed's
autolinks syntax. To render the name of the interface along with the name of
the attribute or method, the "visible text" syntax is used (the text after `|`).

The notions of "media error code", official playback position", "poster frame"
and "timeline offset" are not HTMLMediaElement attributes but rather internal
HTML concepts whose definitions are not currently exported. Longer term, we
should make sure that the HTML spec exports these terms (unless we don't need
them anymore).

The text incorrect referred to a `stalled` attribute that does not exist,
instead of to the `stalled` event. Text updated accordingly.