Stop auto-publication to /TR of Open Screen Protocol
The spec is about to be re-published as a Discontinued Draft, so stopping the
auto-publication to /TR.

Publication to the `gh-pages` branch continues, as that may still be useful.
Typically, once the spec gets re-published, the Editor's Draft should also be
updated to reflect the discontinued status and redirect people to the
Application and Network parts.