Add discontinued note, link to /TR, update refs
This adds a note to the Status of This Document section of the previous
Open Screen Protocol to note that the work was discontinued and redirect to
the split documents (I copied the entire section because that makes life
easier for Bikeshed, we shouldn't need to update that document any time soon
in any case).

This also updates the references to IETF documents used in the Application
and Network documents, and adds the links to the /TR versions.